order-service/pyproject.toml
[project]
name = "order-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "cryptography==41.0.7",
    "django==4.2.7",
    "django-cors-headers==4.3.1",
    "django-encrypted-model-fields==0.6.5",
    "djangorestframework==3.14.0",
    "djangorestframework-simplejwt==5.3.0",
    "drf-spectacular==0.27.0",
    "factory-boy==3.3.0",
    "grpcio==1.59.3",
    "grpcio-tools==1.59.3",
    "hypothesis==6.92.1",
    "pybreaker==1.0.2",
    "pytest==7.4.3",
    "pytest-cov==4.1.0",
    "pytest-django==4.7.0",
    "python-dotenv==1.0.0",
    "requests==2.31.0",
    "tenacity==8.2.3",
]
